QCEW_2002_R_Markdown_Master
R Markdown document that allows for reproducible results in bridging and data frame creation in a user's R session. It provides instructions and elaborates on how the program works and its functionality. It is specific to the year 2002, but can be altered within the document and the raw R code.